Email Completed Offers
First draft of email generation.
TODO: There needs to be some thought put into how to query the most
recently completed offers efficiently. Perhaps a new model is required
to keep track of completed offers and ones that have already been sent.
This solution would be highly specialized but it may be the most efficient.

-Seth