[#57703664] - Refactor Knotis Views
Work in progress.

I think this is pretty close to where it needs to be which is when creating a new view you should only have to specify a template and override process_context to populate the context and optionally hook the view up in urls.py and have everything work out of the box.