Delete Profile Images
I implemented an api for the ImageInstance model that handles getting and deleting image instances.
I implemented an api view class to expose the ImageInstanceApi
I implemented a delete x on each image in the profile carousel that is rendered when the current identity is a manager of that business that allows the user to delete the image by clicking on the x.