EmbeddedView and MVP Static Files
Created a new view class called EmbeddedView. The embedded view is a view that is meant to be fetched via an AJAX request and rendered into an existing DOM. The tricky part is when we want to return this view when requested via an external URL. The EmbeddedView handles this case by allowing a parent_view_class to be defined on the class (this can also be overridden via the constructor). In the case where the page is requested from an external URL the view is rendered to a string and added to the context and then rendered to the parent view.

Also added static files required by the new MPV views.

Created a new master template called base.html