[#78751788] - Test Passport Book XAPI
Pretty sure this task is complete. Don't have passport test data in my dev environment so this still needs to be tested.
Refactored the "create_transaction_transfer" method logic from the TransactionManger to the TransactionApi.
Renamed TransactionApi "create_transaction_transfer" to "transfer_transaction_collection"
Removed terrible TransactionManager create method that was not being consumed which is good because the behavior of the implementation was wildly unpredictible (and almost always certianly wrong)