[#78751788] - Test Passport Book XAPI
Passport book connect and redeem methods now only return the transaction that belongs to the current identity.
Passport book redeem method now checks for a valid page number parameter.
The test for wether a passport book has been transfered or not now raises an exception in the case where there is no transaction collection items attached to a transaction collection.