[Finish #79980004] - Follow Button No Longer Opens Up Authentication Modal When Logged Out
So the problem is that we replaced the follow button (which was an HTML anchor) with an HTML form. I'm pretty sure the navigation system doesn't try to do any of it's navigation magic with forms so this is causing the response from hitting the submit button on the form from being rendered in a modal.

I took a crack at the logic for rendering an anchor tag instead of the form but I'm not 100% satisfied with my approach as there are a couple things wrong with it:

    1. The action button was intended to be a dropdown or popup menu hence the for loop in the template. However somewhere down the line this alternative logic has been replaced with just looping over and creating button after button for each action item. This displeases me.

    2. I'm not sure that the use of request.user.is_authenticated in the base action button template is the right choice as there may be cases where we want the action button to do something other then auth if there is no authenticated user even though there is no current use case where that is true.

After modifying the template to render an anchor tag when logged out instead of the follow form the desired behavior has returned and the auth modal pops up when logged out. The behavior is currently correct but the implementation leaves a little to be desired. Works for the case we are using it in but will probably have to be revisited in the future when we want to use this for more complex use cases.

Also caught a couple PEP-8 Issues