[#80047308] - Paging On Merchant Redemption View
I've started this task by creating a new view mixin in the base knotis.views package. This mixin understands how paging works as far as retrieving the appropriate parameters from a RequestContext object and using these parameters to slice a fully configurable queryset.

I refactored the RedempionGrid class to utilize this mixin as well as the GetCurrentIdentityMixin. I also hooked up the RedemptionGrid view to a url pattern defined in the knotis.contrib.merchant.urls module.

Also fixed up some PEP-8 issues.

This does not include the client side javascript to support paging. That will be coming in a latter pull request.