[Finished #80257100] - Tile Text Sizing
Resized the font and adjusted the margin of the h1 tag on the offer tiles for the mobile (less than 400px) moved the ellipsis from 2 lines to 3 on the mobile and 2 to 4 on desktop. This will allow for larger titles to display, but still handle extremely long titles that can't fit on the tile.