[Finished #79337896] -  Forgot Password Modal Needs Styles
Styled the forgot password modal similar to the create business modal.
I added a "header" block to the forgot.html file to get the title in the top of the modal, I'm not sure if this is how you want me to add the title or if there is another way to do this.