[Fixes #81035626] - Navigation Error When Going To Passport View
So good thing to know: When you load in script tags via an ajax get the $(document).onReady event never gets triggered so you kind of need to have your jquery files do something like:

    var onReady = function () {};
    if ($.isReady) {
        onReady();
    } else {
        $(function() { onReady(); });
    }

This will make sure that your javascript always gets run in all cases.

Changing to this design pattern fixed a lot of issues on the passport page (including purchasing passports)

I also modified the stripe_form.js to check if the stripe js has loaded yet, if not, it tries again in 10ms.

The last change I made was to remove the $.ajaxSetup crossDomain: false option as we want jquery to determine if it's cross domain or not and handle this appropriately.