[Finish #81427756] Added process_context to base email view.
Some may argue this is a bad idea but in general, we'll have a number of
things in our emails. One of which will definitely be something pointing
to knotis.com (```BASE_URL```) and some static material like images
(```STATIC_URL_ABSOLUTE```). This means that any email sent with the EmailView
from ```knotis/views/views.py``` will have the static url and base url
pointing to knotis.com.  I can imagine a scenerio where this is a faulty
assumption - say we want to allow merchants to send emails that don't
actually point to knotis in anyway.

Anyway, this should be fine for a few years.