[Finished #81726102] - Passport Connect Page
Styled the connect passport page for both logged_in users and users that are not logged in.
Simley helped me hack the qrcode view.py file (which I will revert after this), but I never had the logic for logged_in working, so that may work with the real view.