[Finished #82420022] - Re-Style Random Pin and Redemption Voucher Modal
Removed all the old information that was displayed on the modal and styled close button and header section of the modal.
note: Button on the bottom of the modal seems to have gone back to some old styles that make it 50% width on mobile devices, I made the .modal-cta class for the container of the close button, so that I can add the styles more easily to other modals.