[Finish #84162580, #84165642] Fix action button on profile page.
Also decouple some of the css from away screen.css because that shit is
busted.

The establishment profile action button was given the wrong type of
context which caused the action button to get rendered as an anchor tag
rather than a form.

This then broke some of the CSS and so I tag teamed that with my boi
mdgcat.