[Finish #93946066] - CQL Remote Import Command
The CQL Remote Import Commmand executes a remote export of data in a cassandra cluster and imports it into the current database configured by your Django project.

The script supports both CQL3 and Thrift clients and allows you to include or exclude models and allows you to limit the number of rows you want to import.

Also:
    * Removed initial migration files.
    * Fixed an issue in a couple places where UUID's weren't getting properly converted to strings.